+++ lib/firebase_cached_image.dart
@@ -3,6 +3,6 @@ export 'package:firebase_cached_image/src/core/firebase_url.dart';
 export 'package:firebase_cached_image/src/core/image_not_found_exception.dart';
 export 'package:firebase_cached_image/src/core/source.dart';
 export 'package:firebase_cached_image/src/firebase_cache_manager/stub_firebase_cache_manager.dart'
-    if (dart.library.html) "package:firebase_cached_image/src/firebase_cache_manager/web_firebase_cache_manager.dart"
+    if (dart.library.js_interop) "package:firebase_cached_image/src/firebase_cache_manager/web_firebase_cache_manager.dart"
     if (dart.library.io) "package:firebase_cached_image/src/firebase_cache_manager/mobile_firebase_cache_manager.dart";
 export 'package:firebase_cached_image/src/firebase_image_provider.dart';